+++ piano-analytics/src/main/java/io/piano/android/analytics/DeviceInfoProvider.kt
@@ -8,7 +8,6 @@ import android.net.ConnectivityManager
 import android.net.NetworkCapabilities
 import android.net.NetworkInfo
 import android.os.Build
-import android.telephony.TelephonyManager
 import android.util.DisplayMetrics
 import androidx.core.content.ContextCompat
 import io.piano.android.analytics.model.ConnectionType
@@ -65,38 +64,6 @@ internal class DeviceInfoProvider(
     private fun NetworkInfo?.toConnectionType(): ConnectionType = when {
         this == null || !isConnected -> ConnectionType.OFFLINE
         type == ConnectivityManager.TYPE_WIFI -> ConnectionType.WIFI
-        else -> ContextCompat.getSystemService(context, TelephonyManager::class.java)?.run {
-            val type = if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.N) dataNetworkType else networkType
-            type.toConnectionType()
-        } ?: ConnectionType.UNKNOWN
-    }
-
-    private fun Int.toConnectionType(): ConnectionType = when (this) {
-        TelephonyManager.NETWORK_TYPE_GPRS -> ConnectionType.GPRS
-        TelephonyManager.NETWORK_TYPE_EDGE -> ConnectionType.EDGE
-
-        TelephonyManager.NETWORK_TYPE_CDMA,
-        TelephonyManager.NETWORK_TYPE_IDEN,
-        TelephonyManager.NETWORK_TYPE_1xRTT,
-        -> ConnectionType.TWOG
-
-        TelephonyManager.NETWORK_TYPE_UMTS,
-        TelephonyManager.NETWORK_TYPE_EVDO_0,
-        TelephonyManager.NETWORK_TYPE_EVDO_A,
-        TelephonyManager.NETWORK_TYPE_EVDO_B,
-        -> ConnectionType.THREEG
-
-        TelephonyManager.NETWORK_TYPE_HSPA,
-        TelephonyManager.NETWORK_TYPE_HSDPA,
-        TelephonyManager.NETWORK_TYPE_HSUPA,
-        -> ConnectionType.THREEGPLUS
-
-        TelephonyManager.NETWORK_TYPE_HSPAP,
-        TelephonyManager.NETWORK_TYPE_LTE,
-        -> ConnectionType.FOURG
-
-        TelephonyManager.NETWORK_TYPE_NR -> ConnectionType.FIVEG
-
-        else -> ConnectionType.UNKNOWN
+        else -> ConnectionType.MOBILE
     }
 }