+++ piano-analytics/src/main/java/io/piano/android/analytics/PrivacyModesStorage.kt
@@ -1,16 +1,23 @@
 package io.piano.android.analytics
 
 import io.piano.android.analytics.model.PrivacyMode
+import io.piano.android.analytics.model.PrivacyMode.Companion.toPrivacyMode
 import io.piano.android.analytics.model.PrivacyStorageFeature
+import io.piano.android.consents.PianoConsents
+import io.piano.android.consents.models.Product
+import timber.log.Timber
 import java.util.concurrent.TimeUnit
 
 /**
  * Stores information about privacy modes
  */
-class PrivacyModesStorage internal constructor(
+@Deprecated("Use `PianoConsents` for managing consents instead")
+public class PrivacyModesStorage internal constructor(
     private val configuration: Configuration,
     private val prefsStorage: PrefsStorage,
+    private val pianoConsents: PianoConsents?,
 ) {
+    private val consentsEnabled = pianoConsents?.consentConfiguration?.requireConsent == true
     init {
         prefsStorage.privacyStorageFilter = ::isFeatureAllowed
     }
@@ -26,20 +33,26 @@ class PrivacyModesStorage internal constructor(
      * All registered privacy modes. Add a [PrivacyMode] instance into [allModes] for registering it
      */
     @Suppress("unused", "MemberVisibilityCanBePrivate") // Public API.
-    val allModes = mutableSetOf(
+    public val allModes: MutableSet<PrivacyMode> = mutableSetOf(
         PrivacyMode.NO_CONSENT,
         PrivacyMode.NO_STORAGE,
         PrivacyMode.OPTIN,
         PrivacyMode.OPTOUT,
-        PrivacyMode.EXEMPT
+        PrivacyMode.EXEMPT,
+        PrivacyMode.CUSTOM
     )
 
     /**
      * Current privacy visitor mode
      */
     @Suppress("unused", "MemberVisibilityCanBePrivate") // Public API.
-    var currentMode: PrivacyMode = configuration.defaultPrivacyMode
+    public var currentMode: PrivacyMode = configuration.defaultPrivacyMode
         get() {
+            if (consentsEnabled) {
+                return pianoConsents?.let {
+                    it.consents[it.productsToPurposesMapping[Product.PA]]?.mode
+                }?.toPrivacyMode() ?: PrivacyMode.NO_CONSENT
+            }
             if (field != PrivacyMode.NO_CONSENT && field != PrivacyMode.NO_STORAGE) {
                 if (prefsStorage.privacyExpirationTimestamp in 1..System.currentTimeMillis()) {
                     currentMode = configuration.defaultPrivacyMode
@@ -53,6 +66,9 @@ class PrivacyModesStorage internal constructor(
             return field
         }
         set(value) {
+            if (consentsEnabled) {
+                Timber.w("Calling deprecated privacy mode setter while consents enabled")
+            }
             require(value in allModes) {
                 "Privacy mode ${value.visitorMode} is not registered."
             }
@@ -69,6 +85,9 @@ class PrivacyModesStorage internal constructor(
         System.currentTimeMillis() + TimeUnit.DAYS.toMillis(configuration.privacyStorageLifetime.toLong())
 
     internal fun updatePrefs(mode: PrivacyMode) {
+        if (consentsEnabled) {
+            return
+        }
         prefsStorage.apply {
             privacyMode = mode.visitorMode
             privacyExpirationTimestamp = getNewExpirationTimestamp()
@@ -81,7 +100,7 @@ class PrivacyModesStorage internal constructor(
         }
     }
 
-    companion object {
+    private companion object {
         @JvmStatic
         private val cachedPrivacyStorageFeatures = enumValues<PrivacyStorageFeature>().filter {
             it != PrivacyStorageFeature.ALL